particle: improve error handling in on-click handler
* Make sure to _exit() in all forked code paths that doesn't end in a
  successful exec()
* Report errno to parent process on failure so that we can log it

The second point is implemented by a) manually double forking instead
of calling daemon(), and b) using a pipe in the inner fork to be able
to wait for the child process to either succeed in its exec() call, or
write errno to the pipe on failure.

This fixes hangs and crashes when e.g. the binary we're trying to
execute doesn't exist.