font: force a pixel fixup factor for scalable bitmap fonts
That is, for color bitmap fonts (typically emoji fonts). Use the same
calculation used by fontconfig's 10-scale-bitmap-fonts.conf.

Note: if the user has enabled bitmap font scaling in general, we'll
pick it up and scale the final glyph image, regardless of whether the
font is a regular bitmap font, or a color bitmap font.

But, if the user has disabled bitmap font scaling, we only force a
fixup factor on color bitmap fonts.