module/i3: window event: use 'app_id' for 'application', if available
We previously used the 'pid' property of the 'container' object in a
window event to lookup the application name under /proc/<pid>/comm.

In many cases, the application name is already available in the window
event itself, in the 'app_id' property.

Use this, if available, and fallback to the old PID method otherwise.

A quick experiment suggests 'app_id' is set for all Wayland native
applications, but not for X11 applications running under XWayland.